requirements.txt
gunicorn==19.7.1
httpbin==0.6.2
mock==2.0.0
psycopg2==2.7.4
pytest==3.4.2
requests==2.18.4
SQLAlchemy==1.2.5
weather-api==0.0.6